Tree node interface

MarkdownAST.Node

Base.getindex

Base.push!

Base.pushfirst!

Base.setindex!

MarkdownAST.children

MarkdownAST.haschildren

MarkdownAST.insert_after!

MarkdownAST.insert_before!

MarkdownAST.meta

MarkdownAST.meta!

MarkdownAST.next

MarkdownAST.parent

MarkdownAST.previous

MarkdownAST.unlink!

MarkdownAST.Node — Type

mutable struct Node{M}

Implements a linked list type representation of a Markdown abstract syntax tree, where each node contains pointers to the children and parent nodes, to make it possible to easily traverse the whole tree in any direction. The exact nature of each node in the AST is represented with an instance of some AbstractElement subtype (see: getindex, setindex!, and the Node type acts as a container that links the different nodes together.

Optionally, each node can also store additional meta information, which will be an object of type M (see also meta, meta!). By default, the node does not contain any extra meta information and M = Nothing.

Constructors

Node(c :: AbstractElement)

Constructs a simple standalone node (not part of any tree) without any additional metadata (M = Nothing) containing the Markdown AST element c.

Node{M}(c :: AbstractElement, meta :: M)

Constructs a simple standalone node (not part of any tree) with the meta information meta, containing the Markdown AST element c.

Extended help

The fields of a Node object should not be accessed directly. Rather, there are various methods to query and modify a Markdown AST:

Querying information about the node: haschildren

For accessing neighboring nodes: next, previous, parent, children

To add new nodes as children: push!, pushfirst!, insert_after!, insert_before!

To access or modify the AbstractElement corresponding to a node, use the node[] syntax: getindex, setindex!

To access or modify the extra metadata: meta, meta!

Base.getindex — Method

Base.getindex(node::Node) -> AbstractElement

Returns the Markdown AST element corresponding to node in the AST.

Base.push! — Method

Base.push!(node::Node, child::Node) -> Node

Adds child as the last child node of node. If child is part of another tree, then it is unlinked from that tree first (see unlink!). Returns the parent node.

Base.pushfirst! — Method

Base.pushfirst!(node::Node, child::Node) -> Node

Adds child as the first child node of node. If child is part of another tree, then it is unlinked from that tree first (see unlink!). Returns the parent node.

Base.setindex! — Method

Base.setindex!(node::Node, c::AbstractElement) -> AbstractElement

Updates the Markdown AST element corresponding to node in the AST, setting it to c. Returns the new element c.

MarkdownAST.children — Method

children(node::Node)

Returns an iterator that runs, in sequence, over all the immediate children of the node.

MarkdownAST.haschildren — Method

haschildren(node::Node) -> Bool

Returns true if node has any children nodes and false otherwise.

MarkdownAST.insert_after! — Method

insert_after!(node::Node, sibling::Node) -> Node

Inserts a new child node sibling as the next child after node. node must not be a root node. If sibling is part of another tree, then it is unlinked from that tree first (see unlink!). Returns the original reference node.

MarkdownAST.insert_before! — Method

insert_before!(node::Node, sibling::Node) -> Node

Inserts a new child node sibling as the child right before node. node must not be a root node. If sibling is part of another tree, then it is unlinked from that tree first (see unlink!). Returns the original reference node.

MarkdownAST.meta! — Method

meta!(node::Node{M}, meta::M) -> M

Sets the extra meta information of node to meta, and returns the new meta object.

MarkdownAST.meta — Method

meta(node::Node{M}) -> M

Returns the object storing the extra meta information of the node.

MarkdownAST.next — Method

next(node::Node) -> Union{Node, Nothing}

Returns the next node on the same level, or nothing if this is the last node on this level.

See also: previous.

MarkdownAST.parent — Method

parent(node::Node) -> Union{Node, Nothing}

Returns the parent node, or nothing if node is a root node.

MarkdownAST.previous — Method

previous(node::Node) -> Union{Node, Nothing}

Returns the previous node on the same level, or nothing if this is the first node on this level.

See also: next.

MarkdownAST.unlink! — Method

unlink!(node::Node) -> Node

Isolates and removes the node from the tree by removing all of its links to its neighboring nodes. Returns the updated node, which is now a single, isolate root node.