Markdown AST elements

MarkdownAST.AbstractBlock

MarkdownAST.AbstractElement

MarkdownAST.AbstractInline

MarkdownAST.Admonition

MarkdownAST.BlockQuote

MarkdownAST.Code

MarkdownAST.CodeBlock

MarkdownAST.DisplayMath

MarkdownAST.Document

MarkdownAST.Emph

MarkdownAST.FootnoteDefinition

MarkdownAST.FootnoteLink

MarkdownAST.HTMLBlock

MarkdownAST.HTMLInline

MarkdownAST.Heading

MarkdownAST.Image

MarkdownAST.InlineMath

MarkdownAST.Item

MarkdownAST.Link

MarkdownAST.List

MarkdownAST.Paragraph

MarkdownAST.Strong

MarkdownAST.Table

MarkdownAST.TableBody

MarkdownAST.TableCell

MarkdownAST.TableHeader

MarkdownAST.TableRow

MarkdownAST.Text

MarkdownAST.ThematicBreak

MarkdownAST.can_contain

MarkdownAST.isblock

MarkdownAST.iscontainer

MarkdownAST.isinline

MarkdownAST.AbstractBlock — Type

abstract type AbstractBlock <: AbstractElement

Supertype of all Markdown AST block types.

MarkdownAST.AbstractElement — Type

abstract type AbstractElement

A supertype of all Markdown AST element types.

User-defined elements must not directly inherit this type, but either AbstractBlock or AbstractInline instead.

Interface

By default, each element is assumed to be a leaf element that can not contain other elements as children. An iscontainer method can be defined to override this.

can_contain can be overridden to constrain what elements can be the direct children of another node. By default, inline container elements can contain any inline element and block container elements can contain any block element.

Elements that are implemented as mutable structs should probably implement the equality operator (==), to make sure that two different instances that are semantically the same would be considered equal.

MarkdownAST.AbstractInline — Type

abstract type AbstractInline <: AbstractElement

Supertype of all Markdown AST inline types.

MarkdownAST.Admonition — Type

mutable struct Admonition <: AbstractBlock

Fields

.category :: String: admonition category

.title :: String: admonition title

Constructors

Admonition(category :: AbstractString, title :: AbstractString)

MarkdownAST.BlockQuote — Type

struct BlockQuote <: AbstractBlock

A singleton container element representing a block quote. It must contain other block elements as children.

MarkdownAST.Code — Type

mutable struct Code <: AbstractInline

Inline element representing an inline code span.

Fields

.code :: String: raw code

Constructors

Code(code::AbstractString)

MarkdownAST.CodeBlock — Type

mutable struct CodeBlock <: AbstractBlock

A leaf block representing a code block.

Fields

.info :: String: code block info string (e.g. the programming language label)

.code :: String: code content of the block

MarkdownAST.DisplayMath — Type

mutable struct DisplayMath <: AbstractBlock

Leaf block representing a mathematical display equation.

Fields

.math :: String: TeX code of the display equation

Constructors

DisplayMath(math :: AbstractString)

MarkdownAST.Document — Type

struct Document <: AbstractBlock

Singleton top-level element of a Markdown document.

MarkdownAST.Emph — Type

struct Emph <: AbstractInline

Inline singleton element for emphasis (e.g. italic) styling.

MarkdownAST.FootnoteDefinition — Type

mutable struct FootnoteDefinition <: AbstractBlock

Container block representing the definition of a footnote, containing the definitions of the footnote as children.

Fields

.id :: String: label of the footnote

Constructors

FootnoteDefinition(id :: AbstractString)

MarkdownAST.FootnoteLink — Type

mutable struct FootnoteLink <: AbstractInline

Inline leaf element representing a link to a footnote.

Fields

.id :: String: label of the footnote

Constructors

FootnoteLink(id :: AbstractString)

MarkdownAST.HTMLBlock — Type

mutable struct HTMLBlock <: AbstractBlock

A leaf block representing raw HTML.

MarkdownAST.HTMLInline — Type

mutable struct HTMLInline <: AbstractInline

Inline leaf element representing raw inline HTML.

Fields

.html :: String: inline raw HTML

Constructors

HTMLInline(html::AbstractString)

MarkdownAST.Heading — Type

mutable struct Heading <: AbstractBlock

Represents a heading of a specific level. Can only contain inline elements as children.

Fields

.level :: Int: the level of the heading, must be between 1 and 6.

Constructors

Heading(level :: Integer)

MarkdownAST.Image — Type

mutable struct Image <: AbstractInline

Inline element representing a link to an image. Can contain other inline nodes that will represent the image description.

Fields

.destination :: String: destination URL

.title :: String: title attribute of the link

Constructors

Link(destination::AbstractString, title::AbstractString)

MarkdownAST.InlineMath — Type

mutable struct InlineMath <: AbstractInline

Leaf inline element representing an inline mathematical expression.

Fields

.math :: String: TeX code for the inline equation

Constructors

InlineMath(math::String)

MarkdownAST.Item — Type

struct Item <: AbstractBlock

Singleton container representing the items of a List.

MarkdownAST.Link — Type

mutable struct Link <: AbstractInline

Inline element representing a link. Can contain other inline nodes, but should not contain other Links.

Fields

.destination :: String: destination URL

.title :: String: title attribute of the link

Constructors

Link(destination::AbstractString, title::AbstractString)

MarkdownAST.List — Type

mutable struct List <: AbstractBlock

Represents a Markdown list. The children of a List should only be Items, representing individual list items.

Fields

.type :: Symbol: determines if this is an ordered (:ordered) or an unordered (:bullet) list.

.tight :: Bool: determines if the list should be rendered tight or loose.

Constructors

julia List(type :: Symbol, tight :: Bool)`

MarkdownAST.Paragraph — Type

struct Paragraph <: AbstractBlock

Singleton container representing a paragraph, containing only inline nodes.

MarkdownAST.Strong — Type

struct Strong <: AbstractInline

Inline singleton element for strong (e.g. bold) styling.

MarkdownAST.Table — Type

mutable struct Table <: TableComponent

Container block representing a table, an extension of the CommonMark spec, and should be interpreted as a rectangular grid of cells with a fixed number of rows and columns.

A Table node can only contain either TableHeader or TableBody nodes as children.

TableHeader and TableBody can only contain TableRows as children. A TableHeader should contain only a single TableRow, and any additional ones should be ignored.

Each TableRow contains only TableCells as children. The row with the largest number of cells determines the width (number of columns) of the table.

Since we can not constrain e.g. the number of children or in what order child nodes can appear in a Markdown tree, it is possible to construct tables that can be difficult to interpret. The following rules should be followed when interpreting tables:

The decendants of a Table node should be exactly be a single TableHeader followed by a TableBody.

If the first child is a TableBody, the header row is assumed to be a list of empty cells.

The rows from any nodes following the first TableBody are interpreted as additional table body rows, even if they are contained in a TableHeader.

A Table with no children is interpreted as a table with a single empty header cell.

A TableHeader that is the first child of a Table should only contain one TableRow.

If a TableHeader that is the first child of a Table contains additional rows, the additional rows are interpreted to be body rows.

If a TableHeader that is the first child of a Table is empty, it is assumed to represent a header row with empty cells.

Each TableRow of a table should contain the same number of TableCell.

Any row that has fewer cells than the longest row should be interpreted as if it is padded with additional empty cells.

MarkdownAST.TableBody — Type

struct TableBody <: TableComponent

Represents the body of a Markdown table and should only occur as the second child of a Table node. See Table for information on how to handle other circumstances.

It can only contain TableRow elements.

MarkdownAST.TableCell — Type

mutable struct TableCell <: TableComponent

Represents a single cell in a Markdown table. Can contain inline nodes.

align :: Symbol: declares the alignment of a cell (can be :left, :right, or :center), and should match the .spec field of the ancestor Table

header :: Bool: true if the cell is part of a header row, and should only be true if the cell belongs to a row that is the first

column :: Int: the column index of the cell, which should match its position in the Markdown tree

It is possible that the fields of TableCell are inconsistent with the real structure of the Markdown tree, in which case the structure or the .spec field should take precedence when interpreting the elements.

MarkdownAST.TableHeader — Type

struct TableHeader <: TableComponent

Represents the header portion of a Markdown table and should only occur as the first child of a Table node and should only contain a single TableRow as a child. See Table for information on how to handle other circumstances.

It can only contain TableRow elements.

MarkdownAST.TableRow — Type

struct TableRow <: TableComponent

Represents a row of a Markdown table. Can only contain TableCells as children.

MarkdownAST.Text — Type

mutable struct Text <: AbstractInline

Inline leaf element representing a simply a span of text.

MarkdownAST.ThematicBreak — Type

struct ThematicBreak <: AbstractBlock

A singleton leaf element representing a thematic break (often rendered as a horizontal rule).

MarkdownAST.can_contain — Function

can_contain(parent::AbstractElement, child::AbstractElement) -> Bool

Determines if the child element can be a direct child of the parent element.

This is used to constrain the types of valid children for some elements, such as for the elements that are only allowed to have inline child elements or to make sure that Lists only contain Items.

If the parent element is a leaf node (iscontainer(parent) === false)

Extended help

When extending can_contain for custom abstract classes Markdown elements, similar to the AbstractBlock or AbstractInline elements classes, the second argument to can_contain should always be constrained exactly to ::AbstractElement, in order to avoid method ambiguities. I.e. for some abstract type AbstractT <: AbstractElement, the method should be defined as

can_contain(parent::AbstractT, child::AbstractElement) = ...

For concrete parent types T, where the first argument is constrained as parent::T it should then be fine to take advantage of multiple dispatch when implementing can_contain.

MarkdownAST.isblock — Function

isblock(element::AbstractElement) -> Bool

Determines if element is a block element (a subtype of AbstractBlock).

MarkdownAST.iscontainer — Function

iscontainer(::T) where {T <: AbstractElement} -> Bool

Determines if the particular Markdown element is a container, meaning that is can contain child nodes. Adding child nodes to non-container (leaf) nodes is prohibited.

By default, each user-defined element is assumed to be a leaf node, and each container node should override this method.

MarkdownAST.isinline — Function

isinline(element::AbstractElement) -> Bool

Determines if element is an inline element (a subtype of AbstractInline).